% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_funcs.R
\name{add_func}
\alias{add_func}
\title{Add function to Metodebiblioteket}
\usage{
add_func(func, package, keyword = "r", url = NULL, github = F, update = F)
}
\arguments{
\item{func}{Function name}

\item{package}{Package name}

\item{keyword}{keywords to use seperated by a space}

\item{url}{Url address to help files (if not on CRAN or available on github)}

\item{github}{T or F for if it is a github only package}

\item{update}{T or F on whether to update the function in the dataset}
}
\description{
Add function to Metodebiblioteket
}
