% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_data.R
\docType{data}
\name{aku_testdata}
\alias{aku_testdata}
\title{aku sample testdata}
\format{
An object of class \code{data.frame} with 800 rows and 12 columns.
}
\usage{
data(aku_testdata)
}
\description{
aku sample testdata
}
\keyword{datasets}
