% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_data.R
\docType{data}
\name{aku_testpop}
\alias{aku_testpop}
\title{aku pop testdata}
\format{
An object of class \code{data.frame} with 1600 rows and 7 columns.
}
\usage{
data(aku_testpop)
}
\description{
aku pop testdata
}
\keyword{datasets}
