% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_data.R
\docType{data}
\name{df_empty}
\alias{df_empty}
\title{testdata empty}
\format{
An object of class \code{data.frame} with 20 rows and 2 columns.
}
\usage{
data(df_empty)
}
\description{
testdata empty
}
\keyword{datasets}
