% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_data.R
\docType{data}
\name{df_quartile}
\alias{df_quartile}
\title{testdata quartile}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 30 rows and 5 columns.
}
\usage{
data(df_quartile)
}
\description{
testdata quartile
}
\keyword{datasets}
