% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_data.R
\docType{data}
\name{df_quartile_strata}
\alias{df_quartile_strata}
\title{testdata quartile strata}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 30 rows and 6 columns.
}
\usage{
data(df_quartile_strata)
}
\description{
testdata quartile strata
}
\keyword{datasets}
