% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_data.R
\docType{data}
\name{df_strata}
\alias{df_strata}
\title{testdata strata}
\format{
An object of class \code{data.frame} with 20 rows and 3 columns.
}
\usage{
data(df_strata)
}
\description{
testdata strata
}
\keyword{datasets}
