% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_data.R
\docType{data}
\name{ratioTest}
\alias{ratioTest}
\title{Kostra ratio test data}
\format{
An object of class \code{data.frame} with 34 rows and 3 columns.
}
\usage{
data(kostra_ratioTest)
}
\description{
Kostra ratio test data
}
\keyword{datasets}
