% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{AggrSml2NumVar}
\alias{CalcIndS2}
\alias{CalcInd}
\alias{CalibrateSSB}
\alias{Diff2NumVar}
\alias{EasyData}
\alias{GaussSuppressDec}
\alias{GaussSuppressionFromData}
\alias{Hb}
\alias{HierarchyCompute}
\alias{LmImpute}
\alias{OutlierRegressionMicro}
\alias{PLSroundingPublish}
\alias{PLSrounding}
\alias{PanelEstimation}
\alias{ProtectKostra}
\alias{ProtectTableData}
\alias{Quartile}
\alias{Rank2NumVar}
\alias{SdcForetakPerson}
\alias{SuppressDominantCells}
\alias{SuppressFewContributors}
\alias{SuppressKDisclosure}
\alias{SuppressLinkedTables}
\alias{SuppressSmallCounts}
\alias{SuppressionFromDecimals}
\alias{ThError}
\alias{add_constraint}
\alias{add_outlier}
\alias{confront}
\alias{e.calibrate}
\alias{e.svydesign}
\alias{edit_constraints}
\alias{fill.template}
\alias{get_extremes}
\alias{impute_knn}
\alias{impute_pmm}
\alias{impute_proxy}
\alias{impute_rhd}
\alias{lm}
\alias{make_paramfile}
\alias{model_aggregate}
\alias{modelling_context}
\alias{modifier}
\alias{national_calendar}
\alias{pop.template}
\alias{quantile_weighted}
\alias{sdc_lonn}
\alias{set_arima}
\alias{set_automodel}
\alias{set_outlier}
\alias{set_tradingdays}
\alias{set_transform}
\alias{set_x11}
\alias{shortest_path_cppRouting}
\alias{shortest_path_igraph}
\alias{struktur_model}
\alias{svystatL}
\alias{svystatTM}
\alias{tables_by_formulas}
\alias{validator}
\alias{weights}
\alias{x13_pickmdl}
\alias{x13_spec}
\alias{x13_text_frame}
\alias{x13}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{CalibrateSSB}{\code{\link[CalibrateSSB]{CalibrateSSB}}, \code{\link[CalibrateSSB]{PanelEstimation}}}

  \item{GISSB}{\code{\link[GISSB]{shortest_path_cppRouting}}, \code{\link[GISSB]{shortest_path_igraph}}}

  \item{GaussSuppression}{\code{\link[GaussSuppression]{GaussSuppressDec}}, \code{\link[GaussSuppression]{GaussSuppressionFromData}}, \code{\link[GaussSuppression]{SuppressDominantCells}}, \code{\link[GaussSuppression]{SuppressFewContributors}}, \code{\link[GaussSuppression]{SuppressKDisclosure}}, \code{\link[GaussSuppression]{SuppressLinkedTables}}, \code{\link[GaussSuppression]{SuppressSmallCounts}}, \code{\link[GaussSuppression]{SuppressionFromDecimals}}}

  \item{Kostra}{\code{\link[Kostra]{AggrSml2NumVar}}, \code{\link[Kostra]{Diff2NumVar}}, \code{\link[Kostra]{Hb}}, \code{\link[Kostra]{LmImpute}}, \code{\link[Kostra:OutlierRegression]{OutlierRegressionMicro}}, \code{\link[Kostra]{ProtectKostra}}, \code{\link[Kostra]{Quartile}}, \code{\link[Kostra]{Rank2NumVar}}, \code{\link[Kostra]{ThError}}}


  \item{ReGenesees}{\code{\link[ReGenesees]{e.calibrate}}, \code{\link[ReGenesees]{e.svydesign}}, \code{\link[ReGenesees]{fill.template}}, \code{\link[ReGenesees]{pop.template}}, \code{\link[ReGenesees]{svystatL}}, \code{\link[ReGenesees]{svystatTM}}, \code{\link[ReGenesees]{weights}}}

  \item{SSBpris}{\code{\link[SSBpris]{CalcInd}}, \code{\link[SSBpris]{CalcIndS2}}}

  \item{SSBtools}{\code{\link[SSBtools]{HierarchyCompute}}, \code{\link[SSBtools]{model_aggregate}}, \code{\link[SSBtools]{quantile_weighted}}, \code{\link[SSBtools]{tables_by_formulas}}}

  \item{SdcForetakPerson}{\code{\link[SdcForetakPerson]{SdcForetakPerson}}}

  \item{SmallCountRounding}{\code{\link[SmallCountRounding]{PLSrounding}}, \code{\link[SmallCountRounding:PLSrounding]{PLSroundingPublish}}}

  \item{dcmodify}{\code{\link[dcmodify]{modifier}}}

  \item{easySdcTable}{\code{\link[easySdcTable]{EasyData}}, \code{\link[easySdcTable:ProtectTable]{ProtectTableData}}}



  \item{sdclonn}{\code{\link[sdclonn]{sdc_lonn}}}

  \item{simputation}{\code{\link[simputation:impute_hotdeck]{impute_knn}}, \code{\link[simputation:impute_hotdeck]{impute_pmm}}, \code{\link[simputation]{impute_proxy}}, \code{\link[simputation:impute_hotdeck]{impute_rhd}}}

  \item{stats}{\code{\link[stats]{lm}}}

  \item{struktuR}{\code{\link[struktuR]{get_extremes}}, \code{\link[struktuR]{struktur_model}}}

  \item{validate}{\code{\link[validate]{confront}}, \code{\link[validate]{validator}}}

  \item{pickmdl3}{\code{\link[pickmdl3]{add_constraint}}, \code{\link[pickmdl3]{edit_constraints}}, \code{\link[pickmdl3]{make_paramfile}}, \code{\link[pickmdl3]{x13_pickmdl}}, \code{\link[pickmdl3]{x13_text_frame}}}
  \item{rjd3toolkit}{\code{\link[rjd3toolkit]{add_outlier}}, \code{\link[rjd3toolkit]{modelling_context}}, \code{\link[rjd3toolkit]{national_calendar}}, \code{\link[rjd3toolkit]{set_arima}}, \code{\link[rjd3toolkit]{set_automodel}}, \code{\link[rjd3toolkit]{set_outlier}}, \code{\link[rjd3toolkit]{set_tradingdays}}, \code{\link[rjd3toolkit]{set_transform}}}
  \item{rjd3x13}{\code{\link[rjd3x13]{set_x11}}, \code{\link[rjd3x13]{x13}}, \code{\link[rjd3x13]{x13_spec}}}

}}

