% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_data.R
\docType{data}
\name{testdata}
\alias{testdata}
\title{Kostra test data}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 416 rows and 25 columns.
}
\usage{
data(kostra_testdata)
}
\description{
Kostra test data
}
\keyword{datasets}
