% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{EasyData}
\alias{CalibrateSSB}
\alias{PanelEstimation}
\alias{validator}
\alias{confront}
\alias{ThError}
\alias{Hb}
\alias{Quartile}
\alias{OutlierRegressionMicro}
\alias{Rank2NumVar}
\alias{Diff2NumVar}
\alias{AggrSml2NumVar}
\alias{get_extremes}
\alias{impute_proxy}
\alias{impute_knn}
\alias{lm}
\alias{impute_pmm}
\alias{impute_rhd}
\alias{LmImpute}
\alias{HierarchyCompute}
\alias{PLSrounding}
\alias{PLSroundingPublish}
\alias{ProtectKostra}
\alias{ProtectTableData}
\alias{SdcForetakPerson}
\alias{SuppressSmallCounts}
\alias{SuppressDominantCells}
\alias{SuppressFewContributors}
\alias{SuppressKDisclosure}
\alias{GaussSuppressDec}
\alias{GaussSuppressionFromData}
\alias{SuppressionFromDecimals}
\alias{shortest_path_igraph}
\alias{shortest_path_cppRouting}
\alias{CalcInd}
\alias{CalcIndS2}
\alias{struktur_model}
\alias{model_aggregate}
\alias{quantile_weighted}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{CalibrateSSB}{\code{\link[CalibrateSSB]{CalibrateSSB}}, \code{\link[CalibrateSSB]{CalibrateSSB}}, \code{\link[CalibrateSSB]{PanelEstimation}}, \code{\link[CalibrateSSB]{PanelEstimation}}}

  \item{GISSB}{\code{\link[GISSB]{shortest_path_cppRouting}}, \code{\link[GISSB]{shortest_path_igraph}}}

  \item{GaussSuppression}{\code{\link[GaussSuppression]{GaussSuppressDec}}, \code{\link[GaussSuppression]{GaussSuppressionFromData}}, \code{\link[GaussSuppression]{SuppressDominantCells}}, \code{\link[GaussSuppression]{SuppressFewContributors}}, \code{\link[GaussSuppression]{SuppressKDisclosure}}, \code{\link[GaussSuppression]{SuppressSmallCounts}}, \code{\link[GaussSuppression]{SuppressionFromDecimals}}}

  \item{Kostra}{\code{\link[Kostra]{AggrSml2NumVar}}, \code{\link[Kostra]{Diff2NumVar}}, \code{\link[Kostra]{Hb}}, \code{\link[Kostra]{LmImpute}}, \code{\link[Kostra:OutlierRegression]{OutlierRegressionMicro}}, \code{\link[Kostra]{ProtectKostra}}, \code{\link[Kostra]{Quartile}}, \code{\link[Kostra]{Rank2NumVar}}, \code{\link[Kostra]{ThError}}}

  \item{SSBpris}{\code{\link[SSBpris]{CalcInd}}, \code{\link[SSBpris]{CalcIndS2}}}

  \item{SSBtools}{\code{\link[SSBtools]{HierarchyCompute}}, \code{\link[SSBtools]{model_aggregate}}, \code{\link[SSBtools]{quantile_weighted}}}

  \item{SdcForetakPerson}{\code{\link[SdcForetakPerson]{SdcForetakPerson}}}

  \item{SmallCountRounding}{\code{\link[SmallCountRounding]{PLSrounding}}, \code{\link[SmallCountRounding:PLSrounding]{PLSroundingPublish}}, \code{\link[SmallCountRounding:PLSrounding]{PLSroundingPublish}}}

  \item{easySdcTable}{\code{\link[easySdcTable]{EasyData}}, \code{\link[easySdcTable:ProtectTable]{ProtectTableData}}}

  \item{simputation}{\code{\link[simputation:impute_hotdeck]{impute_knn}}, \code{\link[simputation:impute_hotdeck]{impute_pmm}}, \code{\link[simputation]{impute_proxy}}, \code{\link[simputation:impute_hotdeck]{impute_rhd}}}

  \item{stats}{\code{\link[stats]{lm}}}

  \item{struktuR}{\code{\link[struktuR]{get_extremes}}, \code{\link[struktuR]{struktur_model}}}

  \item{validate}{\code{\link[validate]{confront}}, \code{\link[validate]{validator}}}
}}

